adds review display functionality
